split platform-dock-layout
