+++ src/shared/models/docking-framework.model.ts
@@ -1,5 +1,5 @@
 import { MutableRefObject, ReactNode } from 'react';
-import { DockLayout, DropDirection, LayoutBase } from 'rc-dock';
+import { BoxData, DockLayout, DropDirection, LayoutBase, PanelData, TabData } from 'rc-dock';
 import {
   SavedWebViewDefinition,
   WebViewDefinition,
@@ -191,3 +191,20 @@ export type PapiDockLayout = {
    */
   testLayout: LayoutBase;
 };
+
+export type TabType = string;
+
+export type RCDockTabInfo = TabData & TabInfo;
+
+/**
+ * Check if the input item is just a tab, i.e. not a panel, box, or float.
+ *
+ * @param tab To check.
+ * @returns `true` if its a tab or `false` otherwise.
+ */
+export function isTab(tab: PanelData | TabData | BoxData | undefined): tab is TabData {
+  // Assert the more specific type. Null to work with the external API.
+  // eslint-disable-next-line no-type-assertion/no-type-assertion, no-null/no-null
+  if (!tab || (tab as TabData).title == null) return false;
+  return true;
+}
